delete

+++ src/main/scala/com/cpc/spark/ml/recallReport/model_effect.scala
@@ -97,11 +97,11 @@ object model_effect {
 
   def hash(seed:Int, dist_map:mutable.Map[Int, String])= udf {
     x:String => {
-      var hash_value = Murmur3Hash.stringHash32(x,seed)
+      var hash_value = Murmur3Hash.stringHash32(x,seed).toDouble
       if(hash_value<0){
         hash_value += scala.math.pow(2,32)
       }
-      val dis = hash_value%1000
+      val dis = hash_value.toInt%1000
       dist_map(dis)
     }
   }
